+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 
 v1.0.3
 * Image loading from Buffer
+* Input image cropping
 
 v1.0.2 - (11/09/14)
 * Add shift feature for issues with anti-aliasing